# Cached post serialization