--- a/core/server/api/canary/utils/serializers/output/posts.js
+++ b/core/server/api/canary/utils/serializers/output/posts.js
@@ -1,5 +1,7 @@
 const debug = require('ghost-ignition').debug('api:canary:utils:serializers:output:posts');
 const mapper = require('./utils/mapper');
+const moment = require('moment');
+const postsCache = {};
 
 module.exports = {
     all(models, apiConfig, frame) {
@@ -12,7 +14,15 @@ module.exports = {
 
         if (models.meta) {
             frame.response = {
-                posts: models.data.map(model => mapper.mapPost(model, frame)),
+                posts: models.data.map((model) => {
+                    var key = model.id + '_' + JSON.stringify(frame);
+
+                    if (!postsCache[key] || moment(new Date(postsCache[key].updated_at)).diff(new Date(model.updated_at)) < 0) {
+                        postsCache[key] = mapper.mapPost(model, frame);
+                    }
+
+                    return postsCache[key];
+                }),
                 meta: models.meta
             };