@gargol 2019-11-20
@daniellockyer have you checked the memory impact for this way of cashing and maybe how it compares to plain [memoization](https://lodash.com/docs/4.17.15#memoize) of the `mapPost` function?

From an architectural perspective, think it makes sense to introduce a cache across all serializers at once instead of doing it per resource. But first would need to know the impact on the memory usage :)