@daniellockyer 2019-11-21
@gargol I just used [`object-sizeof`](https://github.com/miktam/sizeof) and after the initial serialization step it's reporting `postsCache` = ~74KB for the 7 default posts (~10.6KB per post), which sounds about right considering the length of the keys + values. We could/should hash the `frame` object and append that to the key instead because the keys are currently 391 characters long...

I initially tried checking the heap usage around the serialization step but there is a lot of temporary allocation for `htmlRelativeToAbsolute` and it reports about 550KB per post. Obviously this gets cleaned up during GC, but that's being done each time in the existing code.

I didn't try `_.memoize`, but I agree that this method could be rolled out broader. There's another mapPost call further down the file that isn't cached, and the v2 API could also benefit from it.