@gargol 2020-01-06
@daniellockyer I'd suggest closing down this PR and treating it as a benchmark for what is possible with some very basic caching. Think we should discuss+create an issue describing plan introducing caching approach across whole serialization layer (or another place if there's such need). Wdyt?