@daniellockyer 2020-01-06
@gargol Agree! Closing for now.